requirements.txt
Flask
pandas
python-dotenv
Pillow
numpy